vim: remove plugin for jinja (produces errors and does not that useful)
It produces tons of the following errors:

  Cannot source a directory: "/etc/xdg/nvim/"

This is because of the incorrect syntax here [1]:

    runtime! indent/html.vim		 +setl nosmartindent

It should be something like this:

    runtime! indent/html.vim
    setl nosmartindent

  [1]: https://github.com/lepture/vim-jinja/blob/7c54e2320c5f676636cdd8697480782eebfb2bf2/indent/jinja.vim#L13

However it seems that I don't need this plugin, since it works only with
HTML, but not with SQL, for which I needs it.

And I will ever use it again, I will use Glench/Vim-Jinja2-Syntax
instead, it does not have this bug.