vim: switch to clangd completion
It supports semantic GoTo (I've been written without this feature for a
long time, let's see if it will give something for my development)

But it requires an initial index, that is pretty time consumming, though
completion works just after start, in other words you don't need to make
the index process finished.

On my desktop it takes ~20min to do the initial index, and ~200MiB of
space. So it is not that friendly for laptopts...