vim: documentation preview from right
It is slightly annoying to have this preview window always, especially
in the bottom of the screen. I will be OK with if it was from the right,
but it seems that it is not possible.

So I've just switched to explicit documentation suggest, seems that this
should work for my usecase and, I hope, will be less distractive.